+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 0.4.0
 -----
 
+- Schedulable#tiq deprecation warning removed.
 - Schedules are now stored on the workers directly instead of in a
   pseudo-global, mutable Hash.
 - Clock now start looping automatically if `Sidekiq.server?` returns true.